feature(git): add git configuration to dotfiles
Main git configuration is now stored in dotfiles repo.

Work-specific config is stored in ~/work/.gitconfig to preserve the
privacy and confidentiality of information that may be protected by
company or government policy.

Git config management is currently only enabled on WSL, as more testing
is needed to make sure that nothing is broken on other systems.