fix(zsh): remove fzf plugin from oh-my-zsh
oh-my-zsh would attempt to load the fzf plugin before fzf was on the
path, which would result in an error when the shell was opened. This
error only appeared on macOS because /opt/homebrew/bin is not on the
path until the ~/.zsh_platform file is sourced, while on other systems
fzf is installed to a folder that is already in the path.