feat(scripts): strip home directory from fzf input
Strip the path to the user's home directory out of all items in the
list. This helps when the desired directory contains a substring of the
value of $HOME (e.g. homelab). Since the string 'home' would appear in
every entry (assuming your $HOME is in /home/) then the desired item
will be harder to select in fzf. This also makes the assumption that all
directories will be inside $HOME.